Move dialog to directory
